admin_control

+++ app/controllers/users_controller.rb
@@ -1,5 +1,6 @@
 class UsersController < ApplicationController
   before_action :authenticate_user!
+  after_action  :verify_authorized 
   def index
     @users = User.all
     authorize User
@@ -12,7 +13,25 @@ class UsersController < ApplicationController
 
   def destroy
     user = User.find(params[:id])
+    authorize user
     user.destroy
     redirect_to new_user_session_path, notice: "User_deleted!"
   end
+
+  def update
+    @user = User.find(params[:id])
+    authorize @user
+    if @user.update_attributes(secure_params)
+      redirect_to users_path, success: "User updated"
+    else
+      redirect_to users_path, alert: "Unable to update user"
+    end
+  end
+
+  private
+
+    def secure_params
+      params.require(:user).permit(:role)
+    end 
+
 end
